requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pyserial==3.5
requests==2.32.4
SQLAlchemy==2.0.42
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
